Update README.md
This is an open source but also user-friendly code for surrogate modelling of SMRF considering earthquakes as the dominant hazard. The merit of this code is that users can put some simple input values for the given building and be provided with the maximum inter story drift values as the conventional demand parameter.